pom: Update dependencies to ph-genericode 7.1.1 and JAXB 4.0.4
ph-genericode 7.x uses JAXB 4.0, so we need to update our explicit
dependency.

JAXB 4.0 brings the namespace change from javax.xml.bind to
jakarta.xml.bind. This does not impact us directly, as we don't have any
reference to it, but it might impact users who rely on the indirect
dependency on a JAXB implementation.