pom: Check API changes against SemVer rules during build
Use japicmp to detect any changes in the public API compared to the
previous release, and fail the build if the rules of semantic versioning
are violated.

This is run during the "verify" phase, just before the "install" phase.
It also outputs the API differences in the target/japicmp folder.