SdkComponentDescriptor: Fix equals() and hasCode() methods
We need to also consider the "implType" field when comparing two
instances or calculating the hash code.

This makes the check in SdkComponentFactory that there are two
components with the same version and type actually work and throw the
intended exception.
As we have this situation in unit test, any test that uses
SdkComponentFactory currently fails because of this exception.
This will addressed in the next commit.