SdkComponent: Add optional qualifier field
This qualifier enables having several components implementations for the
same SDK version and with the same component type. You can choose the
one you want by indicating its qualifier.

The qualifier defaults to the empty string, so this is backwards
compatible for existing code. The behaviour will be the same if you don't
have components with the same version and type. but if you do an
exception will be thrown instead of returning arbitrarily one of the
component implementation.

Update and add unit tests to check the expected behaviour.